+++ frontend/app/address/layout.tsx
@@ -7,7 +7,7 @@ export default function Layout({ children }: { children: ReactNode }) {
 
   return (
     <section className="grow flex flex-col justify-around items-center space-y-4">
-      <h1 className={title()}>Conta</h1>
+      <h1 className={title()}>Endereço</h1>
       {children}
     </section>
   );