+++ frontend/next.config.js
@@ -1,6 +1,20 @@
 /** @type {import('next').NextConfig} */
 const nextConfig = {
-  output: "export",
+  env: {
+    BACKEND_URL: "https://pit2-api.nubling.dev",
+  },
+  images: {
+    remotePatterns: [
+      {
+        protocol: "https",
+        hostname: "picsum.photos",
+      },
+      {
+        protocol: "https",
+        hostname: "nubling-dev.s3.amazonaws.com",
+      },
+    ],
+  },
 };
 
 module.exports = nextConfig;