+++ frontend/types/cookies-keys.enum.ts
@@ -1,0 +1,4 @@
+export enum CookiesKeys {
+  accountId = "@pit-cookie-account-id",
+  favoriteAddressId = "@pit-cookie-account-address-id",
+}