+++ CHANGELOG.md
@@ -5,6 +5,37 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.9.0] - 2023-10-14
+
+### Changed
+
+- [Frontend] Versão 0.2.0
+
+### Added
+
+- [Frontend] Home / Landing Page
+- [Frontend] Pagina buscar conta por email (Login)
+- [Frontend] Pagina criar nova conta
+- [Frontend] Pagina exibir dados da conta
+- [Frontend] Pagina criar novo endereço
+- [Frontend] Pagina categorias
+- [Frontend] Pagina cupcakes de uma categoria
+
+## [0.9.0] - 2023-10-14
+
+### Changed
+
+- [Backend] Versão 0.6.1
+- [Backend] Habilitando CORS
+
+## [0.8.0] - 2023-10-08
+
+### Added
+
+- [Frontend] Versão 0.1.0
+- [Frontend] Inicialização do projeto utilizando NextJS e NextUI
+- [CI] Criação das pipelines utilizando github action para criação e publicação das imagens utilizando o Docker Hub
+
 ## [0.7.0] - 2023-09-25
 
 ### Changed